Enhance inventory and item management system
- Updated stage_first_load.json to include new items and properties for the Lamp.
- Added new images for the Lamp and character animations related to the Lamp.
- Introduced BackpackConfig and BackpackVisuals classes to manage item groups and their visual representation.
- Enhanced Inventory class to support multiple item groups and selection logic.
- Updated SaveData to handle new backpack group data.
- Improved CandleStick class to manage light states and interactions.
- Refactored Character class to support new item visuals and interactions.

+++ src/states/stage/Lighting.cpp
@@ -124,7 +124,7 @@ int StageState::CreateStaticLight(Vec2 pos, bool startsLit) {
 }
 
 void StageState::SetLightEnabled(int lightId, bool enabled) {
-    if (lightId >= 0 && lightId < lights.size()) {
-        lights[lightId].enabled = enabled;
+    if (lightId >= 0 && static_cast<size_t>(lightId) < lights.size()) {
+        lights[static_cast<size_t>(lightId)].enabled = enabled;
     }
 }
